# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Система бронирования мест в кафе",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1/",
      "description": "Локальный сервер"
    }
  ],
...
  "paths": {
...
    "/cafes/{cafe_id}/tables": {
      "get": {
        "tags": [
          "Столы"
        ],
        "summary": "Список столов в кафе",
        "description": "Получение списка доступных для бронирования столов в кафе. Для администраторов и менеджеров - все столы (с возможностью выбора), для пользователей - только активные.",
        "operationId": "get_all_tables_cafe__cafe_id__tables_get",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "cafe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Cafe Id"
            },
            "description": "ID кафе"
          },
          {
            "name": "show_active",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "title": "Показывать активные столы?",
              "default": true
            },
            "description": "Показывать все столы в кафе, только активные или только не активные. <br>По умолчанию показывает: <br><ul><li>для пользователя - только активные столы в кафе (всегда, в не зависимости от значения параметра).</li><li>для администратора - все столы в кафе (и активные и не активные)</li><li>для менеджера - активные столы в кафе </li></ul>"
          }
        ],
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TableInfo"
                  },
                  "title": "Возвращает все столы"
                }
              }
            }
          },
          "401": {
            "description": "Неавторизированный пользователь",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "404": {
            "description": "Данные не найдены",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации данных",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Столы"
        ],
        "summary": "Новый стол в кафе",
        "description": "Создает новое стола кафе. Только для администраторов и менеджеров.",
        "operationId": "create_table_cafe__cafe_id__tables_post",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "cafe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Cafe Id"
            },
            "description": "ID кафе"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TableCreate",
                "description": "Данные нового стола"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Успешно",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableInfo"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка в параметрах запроса",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "401": {
            "description": "Неавторизированный пользователь",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "404": {
            "description": "Данные не найдены",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации данных",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          }
        }
      }
    },
    "/cafes/{cafe_id}/tables/{table_id}": {
      "get": {
        "tags": [
          "Столы"
        ],
        "summary": "Информация о столе в кафе по его ID",
        "description": "Получение информации о столе в кафе по его ID. Для администраторов и менеджеров - все столы, для пользователей - только активные.",
        "operationId": "get_table_by_id_cafe__cafe_id__tables__table_id__get",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "cafe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Cafe Id"
            },
            "description": "ID кафе"
          },
          {
            "name": "table_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Table Id"
            },
            "description": "ID стола"
          }
        ],
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableInfo"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка в параметрах запроса",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "401": {
            "description": "Неавторизированный пользователь",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "404": {
            "description": "Данные не найдены",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации данных",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Столы"
        ],
        "summary": "Обновление информации о столе в кафе по его ID",
        "description": "Обновление информации о столе в кафе по его ID. Только для администраторов и менеджеров.",
        "operationId": "path_table_by_id_cafe__cafe_id__tables__table_id__patch",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "cafe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Cafe Id"
            },
            "description": "ID кафе"
          },
          {
            "name": "table_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "ID стола",
              "title": "Table Id"
            },
            "description": "ID стола"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TableUpdate",
                "description": "Обновленные данные стола"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableInfo"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка в параметрах запроса",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "401": {
            "description": "Неавторизированный пользователь",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "404": {
            "description": "Данные не найдены",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации данных",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CafeShortInfo": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "address": {
            "type": "string",
            "title": "Address"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "photo_id": {
            "type": "string",
            "format": "uuid",
            "title": "Photo Id"
          }
        },
        "type": "object",
        "title": "CafeShortInfo"
      },
...
      "CustomError": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "CustomError"
      },
...
      "TableCreate": {
        "properties": {
          "description": {
            "type": "string",
            "title": "Description"
          },
          "seat_number": {
            "type": "integer",
            "title": "Seat Number"
          }
        },
        "type": "object",
        "required": [
          "seat_number"
        ],
        "title": "TableCreate"
      },
      "TableInfo": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "cafe": {
            "$ref": "#/components/schemas/CafeShortInfo"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "seat_number": {
            "type": "integer",
            "title": "Seat Number"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "title": "TableInfo"
      },
...
      "TableUpdate": {
        "properties": {
          "description": {
            "type": "string",
            "title": "Description"
          },
          "seat_number": {
            "type": "integer",
            "title": "Seat Number"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "TableUpdate"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}